+++ packages/camera/camera_platform_interface/lib/src/types/image_file_format.dart
@@ -4,14 +4,12 @@
 
 /// The format in which images should be returned from the camera.
 enum ImageFileFormat {
-  /// It will return the image in JPEG format.
-  ///
-  /// Is the widely used format for images.
+  /// The JPEG format.
   jpeg,
 
-  /// It will return the image in HEIF format.
+  /// The HEIF format.
   ///
-  /// HEIF is a file format name that refers to High Efficiency Image Format (HEIF).
-  /// Will works only iOS 11+.
+  /// HEIF is a file format name that refers to High Efficiency Image Format
+  /// (HEIF). For iOS, this is only supported on versions 11+.
   heif,
 }