+++ packages/camera/camera_platform_interface/CHANGELOG.md
@@ -1,6 +1,6 @@
 ## 2.7.0
 
-* Adds support to HEIF format for iOS.
+* Adds support for setting the image file format. See `CameraPlatform.setImageFileFormat`.
 * Updates minimum supported SDK version to Flutter 3.10/Dart 3.0.
 
 ## 2.6.0